FIX: UI showed wrong error message when deleted a moderator (#35496)
The UI wasn't "synced" with the backend and showed the wrong error
message when trying to delete a moderator.

In the back-end, only admins can't be deleted.
In the front-end, it was tied to a "staff" check.

Ref - https://meta.discourse.org/t/-/386013

+++ frontend/discourse/admin/templates/admin-user/index.gjs
@@ -772,7 +772,11 @@ export default <template>
             />
           {{/if}}
         {{else}}
-          {{@controller.deleteAllPostsExplanation}}
+          {{#if @controller.deleteAllPostsExplanation}}
+            <span class="delete-all-posts-explanation">
+              {{@controller.deleteAllPostsExplanation}}
+            </span>
+          {{/if}}
         {{/if}}
       </div>
     </div>
@@ -988,7 +992,9 @@ export default <template>
       <br />
       <div class="pull-right">
         {{icon "triangle-exclamation"}}
-        {{@controller.deleteExplanation}}
+        <span class="delete-explanation">
+          {{@controller.deleteExplanation}}
+        </span>
       </div>
     {{/if}}
   </section>
